pyproject.toml
[tool.poetry]
name = "yunzai-nonebot"
version = "0.1.0"
description = ""
authors = ["realhuhu <2661467107@qq.com>"]
readme = "README.md"
packages = [{ include = "yunzai_nonebot" }]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.51.3"
googleapis-common-protos = "^1.58.0"
nonebot2 = "^2.0.0rc3"
omegaconf = "^2.3.0"
pip = "^23.0.1"
nonebot-adapter-onebot = "^2.2.1"


[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.51.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true
